+++ source/LDL/Platforms/Windows/WinError.cpp
@@ -40,7 +40,7 @@ const char* WindowError::GetErrorMessage()
         lastError,
         MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT),
         _buffer.data(),
-        _buffer.capacity() + 1,
+        (DWORD)_buffer.capacity() + 1,
         NULL
     );
